+++ HomeAccessPlus/HAPi.swift
@@ -493,7 +493,7 @@ class HAPi {
             
             // Formatting the name of the file to make sure that it is
             // valid for storing on Windows file systems
-            fileName = formatInvalidFileName(fileName)
+            fileName = formatInvalidName(fileName)
             
             logger.debug("Name of file being uploaded: \(fileName)")
             
@@ -571,7 +571,7 @@ class HAPi {
             let pathArray = String(formattedPath).componentsSeparatedByString("/")
             fileName = pathArray.last!
             fileName = fileName.stringByRemovingPercentEncoding!
-            fileName = formatInvalidFileName(fileName)
+            fileName = formatInvalidName(fileName)
             logger.debug("Name of file item being deleted: \(fileName)")
             
             // Setting the tokens that are collected from the login, so the HAP+
@@ -666,8 +666,79 @@ class HAPi {
         }
     }
     
-    /// Checking to make sure that the file name doesn't contain
-    /// any names not allowed by Windows
+    /// Creates a new folder in the currently viewed folder
+    ///
+    /// The user can create a folder in the currently browsed
+    /// to folder from the upload popover, and this function
+    /// calls the relevant HAP+ API to create the folder. It also
+    /// checks to make sure there are no forbidden characters in
+    /// the name of the folder, and converts the current folder
+    /// and new folder name into URL encoded strings, so that the
+    /// HAP+ server can create the folder correctly
+    ///
+    /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
+    /// - since: 0.6.0-alpha
+    /// - version: 1
+    /// - date: 2016-01-23
+    ///
+    /// - parameter currentFolder: Path to the folder that is currently
+    ///                            being shown to the user
+    /// - parameter newFolderName: The name of the new folder that is to
+    ///                            be created
+    func newFolder(currentFolder: String, newFolderName: String, callback:(result: Bool) -> Void) -> Void {
+        // Checking that we still have a connection to the Internet
+        if (checkConnection()) {
+            logger.debug("New folder to be created in location: \(currentFolder)")
+            logger.debug("New folder raw name: \(newFolderName)")
+            
+            // Removing invalid characters from the new folder name
+            let formattedNewFolderName = formatInvalidName(newFolderName)
+            
+            // Replacing the escaped slashes with a forward slash
+            // from the current folder path
+            var folderFormattedPath = currentFolder.stringByReplacingOccurrencesOfString("\\\\", withString: "/")
+            folderFormattedPath = folderFormattedPath.stringByReplacingOccurrencesOfString("\\", withString: "/")
+            
+            // Combining the path of the current folder with the name
+            // of the new folder
+            var fullNewFolderPath = folderFormattedPath + "/" + formattedNewFolderName
+            
+            // Escaping any non-allowed URL characters - see: http://stackoverflow.com/a/24552028
+            fullNewFolderPath = fullNewFolderPath.stringByAddingPercentEncodingWithAllowedCharacters(.URLHostAllowedCharacterSet())!
+            
+            logger.debug("New folder being created in formatted location path: \(fullNewFolderPath)")// Setting the json http content type header, as the HAP+
+            // API expects incomming messages in "xml" or "json"
+            // As the user is logged in, we also need to send the
+            // tokens that are collected from the login, so the HAP+
+            // server knows which user has sent this request
+            let httpHeaders = [
+                "Content-Type": "application/json",
+                "Cookie": "token=" + settings.stringForKey(settingsToken1)! + "; " + settings.stringForKey(settingsToken2Name)! + "=" + settings.stringForKey(settingsToken2)!
+            ]
+            
+            // Connecting to the API to create the new folder
+            logger.debug("Attempting to create the new folder")
+            Alamofire.request(.POST, settings.stringForKey(settingsHAPServer)! + "/api/myfiles/new/" + fullNewFolderPath, headers: httpHeaders, encoding: .JSON)
+                // Parsing the JSON response
+                .responseJSON { response in switch response.result {
+                case .Success:
+                    logger.debug("Response from creating new folder: \(response.data)")
+                    // Letting the callback know we have successfully logged in
+                    callback(result: true)
+                    
+                case .Failure(let error):
+                    logger.warning("Request failed with error: \(error)")
+                    callback(result: false)
+                    }
+            }
+        } else {
+            logger.warning("The connection to the Internet has been lost")
+            callback(result: false)
+        }
+    }
+    
+    /// Checking to make sure that the file or folder name doesn't
+    /// contain any names not allowed by Windows
     ///
     /// Apps allow you to name files whatever you want to, as they
     /// are not under the same restrictions as Windows. However, if
@@ -679,26 +750,43 @@ class HAPi {
     /// is not one of the invalid file names, and if so, append an
     /// underscore "_" to the end of it
     ///
+    /// Since 0.6.0-aplha new folders can also be created via the app,
+    /// and this function is also called to make sure that the name
+    /// of the folder does also not contain any invalid characters
+    ///
     /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
     /// - since: 0.5.0-beta
-    /// - version: 1
-    /// - date: 2016-01-15
+    /// - version: 2
+    /// - date: 2016-01-23
     ///
-    /// - parameter fullFileName: The full file name and extension that
-    ///                           is going to be given to the HAP+ server
-    /// - returns: The file name with invalid file names modified
-    func formatInvalidFileName(fullFileName: String) -> String {
+    /// - parameter fullName: The full name (and extension for files)
+    ///                       that is going to be given to the HAP+
+    ///                       server to create the resource
+    /// - returns: The item name with reserved characters modified
+    func formatInvalidName(fullName: String) -> String {
         // Making sure that the file name doesn't contain any reserved
         // names, which Windows forbids, meaning the file will be
         // inaccessable. See: https://msdn.microsoft.com/en-gb/library/windows/desktop/aa365247(v=vs.85).aspx#naming_conventions
         let reservedNames = ["CON","PRN","AUX","NUL",
             "COM1","COM2","COM3","COM4","COM5","COM6","COM7","COM8","COM9",
             "LPT1","LPT2","LPT3","LPT4","LPT5","LPT6","LPT7","LPT8","LPT9"]
+        let reservedCharacters = ["<",">",":","\"","/","\\","|","?","*"]
+        
+        // Looping around the full name to check that there aren't
+        // any reserved characters in the name. This should be done
+        // before the file name is checked for reserved characters
+        // as any part of the file or folder name should not contain
+        // any reserved characters. Any reserved character that are
+        // found are replaced with an underscore "_"
+        var formattedFullName = fullName
+        for reservedCharacter in reservedCharacters {
+            formattedFullName = formattedFullName.stringByReplacingOccurrencesOfString(reservedCharacter, withString: "_")
+        }
         
         // Invalid file names are in the format <reservedNames>.<ext>
         // but <anything><reservedNames><anything>.<ext> are allowed
         // so we only really need to check if fileName[0] is invalid
-        var fileName = fullFileName.componentsSeparatedByString(".")
+        var fileName = formattedFullName.componentsSeparatedByString(".")
         
         // Looping around each item in the reserved names array to see
         // if fileName[0] matches any of the items
@@ -706,7 +794,7 @@ class HAPi {
             if (reservedName.lowercaseString == fileName[0].lowercaseString) {
                 // An invalid file name has been found, so modify it to
                 // contain an underscore at the end
-                logger.warning("Reserved file name found: \(fullFileName)")
+                logger.warning("Reserved file name found: \(fullName)")
                 fileName[0] = fileName[0] + "_"
             }
         }