+++ HomeAccessPlus/MasterViewController.swift
@@ -389,6 +389,47 @@ class MasterViewController: UITableViewController, UISplitViewControllerDelegate
         })
     }
     
+    /// Creates a new folder inside the currently browsed to folder
+    ///
+    /// This function will be called from the upload popover delegate
+    /// once the user has named the folder. This function will then
+    /// format the folder name, pass it to the HAPi to create the
+    /// folder, then reload the file browser table on successful
+    /// folder creation
+    ///
+    /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
+    /// - since: 0.6.0-alpha
+    /// - version: 2
+    /// - date: 2016-01-23
+    ///
+    /// - parameter folderName: The name of the folder to be created
+    func newFolder(folderName: String) {
+        hudShow("Creating \"" + folderName + "\" folder")
+        logger.debug("Folder name from delegate callback: \(folderName)")
+        
+        // Calling the HAPi to create the new folder
+        api.newFolder(currentPath, newFolderName: folderName, callback: { (result: Bool) -> Void in
+            self.hudHide()
+            
+            // There was a problem with creating the folder, so let the
+            // user know about it
+            if (result == false) {
+                let uploadFailController = UIAlertController(title: "Unable to create folder", message: "The folder was not successfully created. Please check and try again", preferredStyle: UIAlertControllerStyle.Alert)
+                uploadFailController.addAction(UIAlertAction(title: "OK", style: UIAlertActionStyle.Default, handler: nil))
+                self.presentViewController(uploadFailController, animated: true, completion: nil)
+            }
+            
+            // The folder has been created successfuly so we can present and
+            // refresh the current folder to show it
+            if (result == true) {
+                logger.debug("A new folder has been created in: \(self.currentPath)")
+                
+                // Refreshing the file browser table
+                self.loadFileBrowser()
+            }
+        })
+    }
+    
     
     // MARK: MBProgressHUD
     // The following functions look after showing the HUD during the login
@@ -782,7 +823,7 @@ class MasterViewController: UITableViewController, UISplitViewControllerDelegate
         let vc = storyboard.instantiateViewControllerWithIdentifier("fileUploadPopover") as! UploadPopoverTableViewController
         vc.delegate = self
         vc.modalPresentationStyle = UIModalPresentationStyle.Popover
-        vc.preferredContentSize = CGSize(width: 320, height: 320)
+        vc.preferredContentSize = CGSize(width: 320, height: 420)
         if let popover: UIPopoverPresentationController = vc.popoverPresentationController! {
             popover.barButtonItem = sender
             popover.delegate = self